move:#8 api, type 선언 도메인 변경
