+++ src/lib/api/detail.api.ts
@@ -37,21 +37,3 @@ export const fetchGetLikeThisBook = async (
 
   return data[0];
 };
-
-export const fetchGetLikeList = async (userId: string | undefined) => {
-  const res = await fetch(
-    `${SUPABASE_URL}/rest/v1/likes?user_id=eq.${userId}&select=*,book:books(*)`,
-    {
-      method: 'GET',
-      headers: {
-        'Content-Type': 'application/json',
-        Authorization: `Bearer ${API_KEY}`,
-        apikey: API_KEY || '',
-      },
-    }
-  );
-
-  const data = await res.json();
-
-  return data;
-};
